+++ docs/versionhistory.rst
@@ -27,6 +27,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
     type of the value passed to ``task_status.started()``
   - The ``Listener`` class is now covariant in its stream type
   - ``create_memory_object_stream()`` now allows passing only ``item_type``
+  - Object receive streams are now covariant and object send streams are correspondingly
+    contravariant
 - Fixed ``CapacityLimiter`` on the asyncio backend to order waiting tasks in the FIFO
   order (instead of LIFO) (PR by Conor Stevenson)
 - Fixed ``CancelScope.cancel()`` not working on asyncio if called before entering the